create coupon trait
